Move healthcheck controller
